Move commonly used types into plugin.go
